Clean up, fix concurrency issues, and report back
This fixes some issues with concurrent access to the path slice that was fed
in. It cleans up the code significantly by borrowing some constructs from the
rewrite branch (and more of that to come soon). Finally, it reports back all
detected closed polygons over a channel. This channel is closed when the
waitgroup exits.