Implement support for diagonal paths
Mostly, this commit implements support for diagonal paths (both open and
closed). Diagonals are weird because they have some rules associated
with them:

 * You may transition from a vertical arrow into a diagonal path.
 * You may transition from a corner into a diagonal path.
 * You may transition from a diagonal path into another diagonal path, a
 corner, a vertical arrow, a horizontal path, or a vertical path.
 * You may transition from a horizontal or vertical path into a diagonal
 path.

No other diagonal connections are permitted. For example, you may not
transition diagonally from corner to corner, between arrows, etc.

This patch also removes the charType and direction types. I doubt the
former is useful with the current API, and the latter isn't useful with
respect to the char type.