Implement object annotations
This change implements object annotations. These annotations allow users
to "tag" objects with a reference that looks similar to a Markdown link
with reference style. The reference points to a JSON object that
contains options that will be attached to the path of the matching
object. There are two ways to specify a reference.

The first way is to embed a reference inside the object. For example:

```
.------.
|[ref] |
'------'

[ref]: {"fill":"#88d"}
```

The object with the reference `ref` will have the SVG attribute `fill`
set to the value `#88d`. Arbitrary SVG options may be set on the path in
this manner.

By default, both the reference and the label are left in the output.
ASCIIToSVG supports some special attributes; these are attributes
prefixed with `a2s:`. Such attributes will not be inserted into the SVG
path. Currently, two special attributes are implemented: `a2s:label` and
`a2s:delref`.

`a2s:label` changes the text at the point of the label. `a2s:delref`
will remove the reference from rendering if it is set. For example:

```
.------.
|[ref] |
'------'

[ref]: {"fill":"#88d", "a2s:label":"hello", "a2s:delref":1}
```

Will render an SVG containing a box with a background color of `#88d`
and the following un-annotated form:

```
.------.
|hello |
'------'
```

This change also brings in an algorithm to detect the proper color to
use given the contrast of the containing object's fill. This is to
improve accessibility. Objects with a dark fill have white text. Objects
with a light (or no) fill have black text. This currently does not
detect a case where an object has no fill and is logically inside a
containing object with a darker background.

Annotations can also target arbitrary paths. This is intended to be able
to style text and line segments in stable diagrams. An example of such
an annotation is found in the logo in the a2s command.

Tests are added. Copyright years fixed. Unused return values from
io.WriteString are elided entirely.