Fix issues with tags and color calculations
I mistakenly implemented the text color calculation only for reference
text. All text needs to have its color chosen based on the background of
its most specific containing object that has a fill. Since users can set
fill to none, we may need to look backwards. If all enclosing polygons
are transparent, the default is for text to be black.

In doing this, I realized that closed paths should have a default fill
as they did in the PHP version. This default fill color is white.
Specifying a default fill provides a consistent view of the SVG
regardless of the background color.