Delete already-fixed-TODO; fix another.
Only apply the rounded corner hint to corners of the box. This has the
side-effect of fixing the stupid logo quirk that actually makes it
(unexpectedly) look like a lowercase 'a'. For years, I've been bugged
that I didn't know why that was happening. Now that I've fixed it, I'm
sad the logo no longer looks as cool.